Minor progress on fetch statement
Did some more research on fetch statements to see where I was going wrong yesterday. Wanted to get more done today but an unexpected event happened that needed to be dealt with, and now I need to prepare for work. I hope to get this done tomorrow and perhaps test to see if it works, then move on to login.