Closer to solving signup.js
I've recently make a breakthrough in solving signup.js going through various methods to solve it.  I got the system to the point where it can input data into the database but I only realize that after trying to fix it before and thus missed the exact point where everything is running after trying to fix an error message. I'm going to attempt to reverse my steps tomorrow and narrow it down to the point where it works again and hopefully get login set as well.